remove decimals in the distance displays

+++ LiveNationApp/src/main/java/com/livenation/mobile/android/na/ui/fragments/LocationFragment.java
@@ -273,7 +273,7 @@ public class LocationFragment extends LiveNationFragment implements ListView.OnI
             if (actualLocation != null) {
                 float[] result = new float[1];
                 Location.distanceBetween(actualLocation.getLat(), actualLocation.getLng(), city.getLat(), city.getLng(), result);
-                float miles = result[0] / Constants.METERS_IN_A_MILE;
+                int miles = (int) (result[0] / Constants.METERS_IN_A_MILE);
                 distance = String.format(MILES_AWAY, miles);
             }
             holder.getText2().setText(distance);
